Started on redesigning functions and classes
