+++ core/src/old/class/borrow.rs
