+++ core/src/old/value/function/as_func.rs
