+++ macro/src/old/bind.rs
