+++ macro/src/old/bind/attrs.rs
