+++ macro/src/old/bind/class.rs
