+++ macro/src/old/bind/constant.rs
