+++ macro/src/old/bind/function.rs
