+++ macro/src/old/derive/has_refs.rs
