+++ macro/src/old/embed.rs
