+++ sys/Cargo.toml
@@ -2,7 +2,7 @@
 name = "rquickjs-sys"
 version = "0.3.1"
 authors = ["Mees Delzenne <mees.delzenne@gmail.com>"]
-edition = "2018"
+edition = "2021"
 license = "MIT"
 readme = "README.md"
 description = "quickjs bindings for rquickjs"
