+++ sys/quickjs.bind.h
@@ -1,0 +1,11 @@
+// A header which imports the all symbols of the quickjs header but also exports the static atoms.
+
+#include "quickjs.h";
+
+enum {
+    __JS_ATOM_NULL = JS_ATOM_NULL,
+#define DEF(name, str) JS_ATOM_ ## name,
+#include "quickjs-atom.h"
+#undef DEF
+    JS_ATOM_END,
+};
